+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5
+
+* Allowing to specify disabled dates
+* Allowing to specify a minimum and a maximum period
+* Fixing an issue with today's date which was incorrectly detected.
+
 ## 0.0.4
 
 * Updating README.md with image.